overflow! run away!

+++ host-services/builtins/messaging.go
@@ -16,8 +16,8 @@ const (
 	messagingServiceMethodRequest     = "request"
 	messagingServiceMethodRequestMany = "requestMany"
 
-	defaultMessagingRequestTimeout     = time.Millisecond * 750
-	defaultMessagingRequestManyTimeout = time.Second * 3
+	defaultMessagingRequestTimeout     = int64(time.Millisecond * 750)
+	defaultMessagingRequestManyTimeout = int64(time.Second * 3)
 )
 
 type MessagingService struct {
@@ -28,8 +28,8 @@ type MessagingService struct {
 }
 
 type messagingConfig struct {
-	RequestTimeoutMs     int `json:"request_timeout_ms"`
-	RequestManyTimeoutMs int `json:"request_many_timeout_ms"`
+	RequestTimeoutMs     int64 `json:"request_timeout_ms"`
+	RequestManyTimeoutMs int64 `json:"request_many_timeout_ms"`
 }
 
 func NewMessagingService(nc *nats.Conn, log *slog.Logger) (*MessagingService, error) {
@@ -43,8 +43,8 @@ func NewMessagingService(nc *nats.Conn, log *slog.Logger) (*MessagingService, er
 
 func (m *MessagingService) Initialize(config json.RawMessage) error {
 
-	m.config.RequestManyTimeoutMs = int(defaultMessagingRequestManyTimeout)
-	m.config.RequestTimeoutMs = int(defaultMessagingRequestTimeout)
+	m.config.RequestManyTimeoutMs = defaultMessagingRequestManyTimeout
+	m.config.RequestTimeoutMs = defaultMessagingRequestTimeout
 
 	if len(config) > 0 {
 		err := json.Unmarshal(config, &m.config)
@@ -109,7 +109,7 @@ func (m *MessagingService) handleRequest(_, _ string,
 		return hostservices.ServiceResultFail(400, "subject is required"), nil
 	}
 
-	resp, err := m.nc.Request(subject, data, time.Duration(m.config.RequestTimeoutMs*int(time.Millisecond)))
+	resp, err := m.nc.Request(subject, data, time.Duration(m.config.RequestTimeoutMs*int64(time.Millisecond)))
 	if err != nil {
 		m.log.Debug(fmt.Sprintf("failed to send %d-byte request on subject %s: %s", len(data), subject, err.Error()))
 		return hostservices.ServiceResultFail(500, "failed to send request"), nil
